Cleaned the __init__.
