+++ app/__init__.py
@@ -1,50 +1,11 @@
-import os
-import sys
-
-from flask import Flask, render_template, redirect, url_for
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
+
+from app.functions import set_basic_configuration_and_views
 
 app = Flask(__name__)
 app.config.from_object('config')
 
 db = SQLAlchemy(app)
 
-########################
-# Configure Secret Key #
-########################
-def install_secret_key(app, filename='secret_key'):
-    """Configure the SECRET_KEY from a file
-    in the instance directory.
-
-    If the file does not exist, print instructions
-    to create it from a shell with a random key,
-    then exit.
-    """
-    filename = os.path.join(app.instance_path, filename)
-
-    try:
-        app.config['SECRET_KEY'] = open(filename, 'rb').read()
-    except IOError:
-        print('Error: No secret key. Create it with:')
-        full_path = os.path.dirname(filename)
-        if not os.path.isdir(full_path):
-            print('mkdir -p {filename}'.format(filename=full_path))
-        print('head -c 24 /dev/urandom > {filename}'.format(filename=filename))
-        sys.exit(1)
-
-if not app.config['DEBUG']:
-    install_secret_key(app)
-
-@app.errorhandler(404)
-def not_found(error):
-    return render_template('404.html'), 404
-
-from app.users.views import mod as usersModule
-app.register_blueprint(usersModule)
-
-from app.songs.views import mod as songsModule
-app.register_blueprint(songsModule)
-
-@app.route("/")
-def index():
-    return redirect(url_for("songs.home"))
+set_basic_configuration_and_views(app)
